fix gps coords type

+++ src/main/java/com/anateam/dto/GpsCoordinatesDto.java
@@ -6,9 +6,9 @@ import io.swagger.v3.oas.annotations.media.Schema;
 
 public record GpsCoordinatesDto(
     @Schema(description = "Latitude", example = "42.8746", requiredMode = Schema.RequiredMode.REQUIRED)
-    Double latitude,
+    BigDecimal latitude,
 
     @Schema(description = "Longitude", example = "74.5698", requiredMode = Schema.RequiredMode.REQUIRED)
-    Double longitude
+    BigDecimal longitude
 ) {}
